FA-246 - FastAPI app starter

+++ scripts/aws_big_lambda/convert_sam_json_to_yaml/run_convert_json_to_yaml.sh
@@ -1,14 +1,26 @@
-#!
+#!/bin/sh
 # run_convert_json_to_yaml.sh
+# 2024-03-04 | CR
 #
 # Convert json to yaml string and files
 # Useful to convert ".chalice/deployment/sam.json" to "scripts/aws_big_lambda/template-sam.yml"
 #
-# 2024-03-04
-#
+REPO_BASEDIR="`pwd`"
+cd "`dirname "$0"`" ;
+SCRIPTS_DIR="`pwd`" ;
 python -m venv venv
 . venv/bin/activate
-pip install -r requirements.txt
-python -m convert_json_to_yaml "../../.chalice/deployment/sam.json" "./template.yml"
+if [ ! -f requirements.txt ]; then
+    pip install pyyaml
+    pip freeze > requirements.txt
+else
+    pip install -r requirements.txt
+fi
+python -m convert_json_to_yaml "${REPO_BASEDIR}/.chalice/deployment/sam.json" "${SCRIPTS_DIR}/../template.yml"
+echo ""
+echo "Converted file in:"
+echo "${SCRIPTS_DIR}/../template.yml"
+echo ""
 deactivate
 rm -rf __pycache__
+rm -rf venv
